Avoid Promise constructor anti-pattern in README
There is no reason to wrap the promise during this example. At least it also is an anti-pattern as described here: https://github.com/petkaantonov/bluebird/wiki/Promise-anti-patterns#the-deferred-anti-pattern
The only difference between the proposed code, and the existing code is, that the old one returns an `Promise<void>` and the new one returns a `Promise<boolean>`, which makes no difference in this particular case.