feat: add configuration and logic for excluding token updates on http requests based on headers
A new type for excluded headers has been added, each instance allows for a header and a list of
values to be associated with it. When defined, each request made will be checked against that list
to see if any of the headers match the items in the list. If there is a match, instead of making
the call to update the token, it simply returns and makes the request. This will allow applications
that have polling endpoints that do need authentication to time out user connections if a user
steps away from the computer for a bit. When left undefined, existing functionality persists, with
the call to check and update the token happening every time a request is made.

Related issue: https://github.com/mauriciovigolo/keycloak-angular/issues/342